Adapt       - BUG FIX: initialize face_level_[] with               face_level_unknown and check before setting to set               instead of min or max
            - BUG FIX: add call to initialize() to update Enzo
              BaryonFields to Cello's after swapping out block_ for
              child_block_ (could avoid via copy)

Debug       - Added TRACE_DEBUG to comm_CommBlock.hpp

Performance - Trying to get perf_compute to work

Refactor    - Addressed some compiler warnings

Output      - BUG FIX (untested) Added time check with time_stop to Schedule

Mesh        - Added Index::bit_string()

Enzo        - Cleaned up initialize() to avoid unecessary loop over
              NumberOfBaryonFields and remove unecessary
              double-initialization

Build       - Reverted to coarsen=0

--HG--
branch : add-refine

+++ src/Cello/field_FieldFace.cpp
@@ -196,8 +196,8 @@ void FieldFace::store (int n, char * array) throw()
 
   size_t index_array = 0;
 
-  const int index_d  = field_descr_->field_id("density");
-  const int index_te = field_descr_->field_id("total_energy");
+  const size_t index_d  = field_descr_->field_id("density");
+  const size_t index_te = field_descr_->field_id("total_energy");
 
   for (size_t index_field=0; index_field<num_fields; index_field++) {
